Initial support for rustc_codegen_gcc
rustc can use different backend and this new script builds all the needed parts
to have GCC as the code generator.

It needs (at the moment) a patched version of GCC and builds it with libjit support.
Some glue code (in Rust) is added in the mix to finally be able to use the rustc
compiler with this dynamic backend.

Fixes https://github.com/compiler-explorer/compiler-explorer/issues/2683