Add clang-rocm 5.2.3 to the docker image
This may need to be expanded in the future. While I would hope that
ROCm HIP would not have a 1-to-1 requirement on the compiler version,
I think the code object manager (comgr) is pretty tightly coupled to
the compiler.

The initial enablement of ROCm HIP library support in Compiler Explorer
is just going to be for the ROCm 5.2.3 runtime, so we'll only need the
ROCm 5.2.3 compiler.