Add AMD ROCm HIP (#47)
* Add AMD ROCm HIP

* Add clang-rocm 5.2.3 to the docker image

This may need to be expanded in the future. While I would hope that
ROCm HIP would not have a 1-to-1 requirement on the compiler version,
I think the code object manager (comgr) is pretty tightly coupled to
the compiler.

The initial enablement of ROCm HIP library support in Compiler Explorer
is just going to be for the ROCm 5.2.3 runtime, so we'll only need the
ROCm 5.2.3 compiler.

* Rename build-hip.sh to build-rocm-hip.sh

There are two variants of the HIP library. HIP can be built for the
NVIDIA CUDA platform or for the AMD ROCm platform. This particular
script builds it for the latter.

* Remove hipamd-rocm-5.2.3-0001-fix-cmake-minimum-version.patch

This patch is useful when building on Ubuntu 20.04 when using the
copy of CMake provided by the distro (CMake 3.16.3). However, the
misc builder docker already installs a much newer copy of CMake,
so this is not needed.

* Remove hipamd-rocm-5.2.3-0002-fix-build-from-tarball.patch

The bug where hipamd could not be built from a tarball was fixed for
ROCm 5.3, but the simpler solution is just building from a git repo.

* Fixup minor errors in script

* Add support for HIP 4.5.2, 5.0.2 and 5.1.3