Initial support for pythran (#79)
The build is a bit atypical. In order to get a
self-sufficient/relocatable "compiler", we use the conda package... But
we cheat a bit by faking its dependencies on a c++ compiler. The conda
package has its own dep and will insteall its own copy of gcc/clang...
But we have plenty of C++ compilers, so all the gcc deps are stubbed.
Pythran knows how to look for g++ using CXX or by relying on PATH.

refs https://github.com/compiler-explorer/compiler-explorer/issues/6079

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>